Update README.md
All analyses performed on death_cancellations file, which is a filtered data set from the original official data set of all voter cancellations in Georgia from 2012-2018. This original data set has been excluded due to the sensitive nature of such data including names, addresses, and other private information of voters, but it is accessible with a special request to the Georgia Secretary of State's Office.